requirements.txt
numpy==1.20.3
scikit-image==0.18.1
PyYAML==5.3.1
tqdm==4.49.0
scikit-learn==0.24.0
pytorch-lightning==1.4.9
opencv-python==4.5.1.48
numba==0.53
imageio
protobuf~=3.19.0
scipy==1.7.3
torchmetrics==0.6.0